:recycle: refactor: OAuth 코드 리팩토링 #10
