+++ server/src/main/java/server/team33/audit/Auditable.java
@@ -16,10 +16,12 @@ public abstract class Auditable {
     private ZonedDateTime createdAt;
     private ZonedDateTime updatedAt;
 
+
     @PrePersist
     public void prePersist(){
         this.createdAt = ZonedDateTime.now();
         this.updatedAt = ZonedDateTime.now();
+
     }
 
     @PreUpdate
