+++ server/src/main/java/server/team33/cart/entity/Cart.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
 import server.team33.audit.Auditable;
 import server.team33.user.entity.User;
