+++ server/src/main/java/server/team33/exception/bussiness/ExceptionCode.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 
 public enum ExceptionCode {
 
+    MEMBER_NOT_FOUND(404,"회원 정보를 찾을 수 없습니다."),
     EXIST_EMAIL(409,"이미 가입한 e-mail입니다."),
     EXIST_DISPLAY_NAME(409,"이미 존재하는 닉네임입니다."),
     EXIST_PHONE_NUMBER(409,"이미 존재하는 연락처입니다."),
