--- server/src/main/java/server/team33/login/config/CorsConfig.java
@@ -1,29 +1,0 @@
-package server.team33.login.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
-
-@Configuration
-public class CorsConfig {
-    @Bean
-    public CorsFilter corsFilter(){
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        CorsConfiguration corsConfiguration = new CorsConfiguration();
-
-        corsConfiguration.setAllowCredentials(true);
-        corsConfiguration.addAllowedOriginPattern("*");
-        corsConfiguration.addAllowedHeader("*");
-        corsConfiguration.addAllowedMethod("*");
-        corsConfiguration.addExposedHeader("Authorization");
-        corsConfiguration.addExposedHeader("Refresh");
-
-        source.registerCorsConfiguration("/**",corsConfiguration);
-
-        return new CorsFilter(source);
-    }
-
-}
