--- server/src/main/java/server/team33/login/dto/LoginDto.java
@@ -1,11 +1,0 @@
-package server.team33.login.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LoginDto {
-    private String username;
-    private String password;
-}
