--- server/src/main/java/server/team33/login/handler/UserAuthenticationEntryPoint.java
@@ -1,38 +1,0 @@
-package server.team33.login.handler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-import server.team33.exception.response.ErrorResponse;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-@RequiredArgsConstructor
-@Slf4j
-@Component
-public class UserAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    @Override
-    public void commence( HttpServletRequest request, HttpServletResponse response, AuthenticationException authException ) throws IOException, ServletException{
-        log.info("AuthenticationEntiryPoint");
-        errorToJson(response, HttpStatus.UNAUTHORIZED);
-    }
-    public static void errorToJson( HttpServletResponse response, HttpStatus status ) throws IOException{
-
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        ErrorResponse exceptions = ErrorResponse.builder() //errorresponse객체에 상태코드와 메시지 주입
-                .status(status.value()).message(status.getReasonPhrase()).build();
-
-        ObjectMapper objectMapper = new ObjectMapper();
-        String errorResponse = objectMapper.writeValueAsString(exceptions); //json형태로 변경
-
-        response.getWriter().write(errorResponse); //바디에 출력
-    }
-
-}
