+++ server/src/main/java/server/team33/user/dto/UserSignUpDto.java
@@ -1,6 +1,5 @@
 package server.team33.user.dto;
 
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -8,7 +7,6 @@ import javax.validation.constraints.NotBlank;
 
 @Getter
 @Setter
-@Builder
 public class UserSignUpDto {
     @NotBlank
     private String email;
@@ -22,7 +20,5 @@ public class UserSignUpDto {
     private String realName;
     @NotBlank
     private String phoneNumber;
-
-
 
 }
