+++ server/src/main/java/server/team33/user/login/details/PrincipalDetails.java
@@ -1,0 +1,75 @@
+package server.team33.user.login.details;
+
+import lombok.Builder;
+import lombok.Data;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import server.team33.user.entity.User;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+
+@Data
+public  class PrincipalDetails implements UserDetails, OAuth2User {
+
+    private User user;
+    private Map<String, Object> attributes;
+
+    @Builder
+    public PrincipalDetails( User user, Map<String, Object> attributes ){
+        this.user = user;
+        this.attributes = attributes;
+    }
+
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities(){
+        List<GrantedAuthority> auth = user.getRoles().stream().map(r -> new SimpleGrantedAuthority("ROLE_" + r)).collect(Collectors.toList());
+        return auth;
+    }
+
+    @Override
+    public String getPassword(){
+        return user.getPassword();
+    }
+
+    @Override
+    public String getUsername(){
+        return user.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired(){
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked(){
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired(){
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled(){
+        return true;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes(){
+        return attributes;
+    }
+
+    @Override
+    public String getName(){
+        return "";
+    }
+}
