+++ server/src/main/java/server/team33/user/login/handler/UserAccessDeniedHandler.java
@@ -1,0 +1,22 @@
+package server.team33.user.login.handler;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Slf4j
+@Component
+public class UserAccessDeniedHandler implements AccessDeniedHandler {
+    @Override
+    public void handle( HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException ) throws IOException, ServletException{
+        log.error("권한 없는 사용자");
+        UserAuthenticationEntryPoint.errorToJson(response, HttpStatus.FORBIDDEN);
+    }
+}
