+++ server/src/main/java/server/team33/user/login/oauth/provider/GoogleUserInfo.java
@@ -1,0 +1,32 @@
+package server.team33.user.login.oauth.provider;
+
+import java.util.Map;
+
+public class GoogleUserInfo implements  OAuth2UserInfo {
+
+    private Map<String,Object> attributes;
+
+    public GoogleUserInfo( Map<String, Object> attributes ){
+        this.attributes = attributes;
+    }
+
+    @Override
+    public String getProviderId(){
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getProvider(){
+        return "google";
+    }
+
+    @Override
+    public String getEmail(){
+        return (String) attributes.get("email");
+    }
+
+    @Override
+    public String getName(){
+        return  (String) attributes.get("name");
+    }
+}
